requirements.txt
typer==0.12.3
beautifulsoup4==4.12.2
requests==2.31.0
rich==13.7.0
anthropic==0.18.0
python-dotenv==1.0.0
